refactor(apache/coyote/*): 패키지 구조 수정 및 test 패키지 도메인과 같게 수정
