+++ tomcat/src/test/java/org/apache/coyote/http11/response/header/ContentTypeTest.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.header;
+package org.apache.coyote.http11.response.header;
 
 import static org.assertj.core.api.Assertions.assertThat;
